requirements.txt
aiogram==2.16
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
frozenlist==1.2.0
idna==3.3
lxml==4.6.4
multidict==5.2.0
Pillow==8.4.0
ply==3.11
pyTelegramBotAPI==4.2.0
pytz==2021.3
requests==2.26.0
slimit==0.8.1
soupsieve==2.2.1
telebot==0.0.4
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.2
